Adoptions API working, deleted country objects, modified app.props to match our schema

+++ Backend/src/main/java/com/it192/backend/repository/AdoptionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AdoptionRepository extends CrudRepository <Adoption, Integer> {
+public interface AdoptionRepository extends CrudRepository <Adoption, Long> {
 }
